Refactor: split Settings registration into Options + Settings_Fields classes
Mirror the ActivityPub plugin's layout. Move the four register_setting()
calls and their sanitize callbacks into a new Atmosphere\Options class;
move the Settings API sections, fields, and their render callbacks into
a new Atmosphere\WP_Admin\Settings_Fields class. Both self-register their
own hooks — admin_init + rest_api_init for Options, load-settings_page_-
atmosphere for Settings_Fields — and are bootstrapped from Atmosphere::-
init(). Admin keeps the admin-only surfaces (menu, OAuth callback,
asset enqueue, reauth notice, admin-post handlers, meta box, REST
client-metadata route).

Updated tests reference the moved methods on their new homes.
